Don't use threading with Qt, instead use QTimer
