+++ glue_wwt/viewer/qt_data_viewer.py
@@ -1,5 +1,7 @@
 from __future__ import absolute_import, division, print_function
 
+from qtpy import QtCore
+
 from glue_qt.viewers.common.data_viewer import DataViewer
 
 from .data_viewer import WWTDataViewerBase
@@ -42,14 +44,14 @@ class WWTQtViewer(WWTDataViewerBase, DataViewer):
         self.set_status('NOTE ON ZOOMING: use the z/x keys to zoom in/out if scrolling does not work')
 
     def __del__(self):
-        self._cleanup()
+        self._cleanup_time_timer()
 
     def _initialize_wwt(self):
         from pywwt.qt import WWTQtClient
         self._wwt = WWTQtClient()
 
     def closeEvent(self, event):
-        self._cleanup()
+        self._cleanup_time_timer()
         self._wwt.widget.close()
         return super(WWTQtViewer, self).closeEvent(event)
 
@@ -57,6 +59,17 @@ class WWTQtViewer(WWTDataViewerBase, DataViewer):
         self.options_widget().setEnabled(True)
         self.layer_view().setEnabled(True)
 
+    # NOTE: Qt needs to use its own QTimer class instead of threading
+
+    def _setup_time_timer(self):
+        self._current_time_timer = QtCore.QTimer()
+        self._current_time_timer.setInterval(1000)
+        self._current_time_timer.timeout.connect(self._update_time)
+        self._current_time_timer.start()
+
+    def _cleanup_time_timer(self):
+        self._current_time_timer.stop()
+        self._current_time_timer = None
 
 # To ensure backward compatibility with old session files, we need to add the
 # Qt viewer to the data_viewer namespace
